small interface refactor
Removes clues.Slice() in place of clues.Values().Slice().
This cuts down on interface bloat, and provies the error
side of the package to provide slices without also adding
a set of err.values-to-slice interfaces.

Also renames Values to In, as the interface of `clues.In()`
reads better than `clues.Values()`.