add stacktrace entry to stacked errors (#26)
Unlike wrap and Stack(err), in the addition of multiple errors via
Stack(e1, e2, .. eN), the initial stack node missed adding a stacktrace
for the line at which it was called, slicing out observation of that
location. This change fixes the missing trace, and adds some additional
testing around stacktrace behavior and formatting.