introduce AddTrace
introduces AddTrace and clues node IDs.  The purpose
of node ids is to produce a branch identifier for log
query aggregation.  For example: if you have an error
with node id `a,b,c,d,e`, you can use that to query
for all logs in the parent process `a,b,c,d` as a way
to expand the log review scope without pulling in
unrelated process logging.

Also includes some fixes for unit tests failing on main.