ensure lowest-level data takes priority (#44)
changes data cascade behavior in errors to prioritize the
earliest-seeded data in all cases. Ex: if an earlier error return
contains a k:v pair, a higher layer wrap on that error should not be
able to replace the value for that key, whether through With() calls or
WC(ctx).

The ultimate result of this is that no caller will need to consider if
they're choosing the "right time" to add a value or embed the context.
It's always going to be correct, with priority favoring the earliest
error.