adds Comments to clues package
Comments provide a space for long-form commenting around a chain of
processes.  Whether as a way to augment the conditions of hitting a
certain process, or as a way to further explain a given error, this
adddion splits the needs of short-form identifiers (such as the error
message) and long form explanation into their own compartments.