introduce Proxy functionality
Ads a new feature to clues: proxies.  A proxy is a way to ensure clues
additions are not lost in cases where downstream or adhoc funcs are
unable to retrieve those values or pass them back through a clues.Err.
By adding a proxy, a higher level instance also receives downstream
additions.