adds NoTrace to err interface
Adds the err.NoTrace() method, which avoids adding any trace information
for the specified error.  This is particularly helpful for removing the
error trace lines for global sentinels.

Also fixes a bug with the new format of error tracing output where funcs
with generics were printing the generics brackets instead of the func
name.