merge err and errWC constructors
back at the beginning of clues, I wanted errors and context to be
completely separate missions.  Ie: orgs could pick one or the other and
be okay.  So at the time I only added clues context with a builder:
clues.New("err").WithClues(ctx).

We ended up using that so much that I made some QoL funcs:
clues.NewWC(ctx, "err").  That helped, but it has also produced a _lot_
of confusion about when it is correct to provide a ctx or not.  For the
longest time I've kept this current design because I didn't want to
demand a context if the erroring function didn't already call for one.

But the complaints about complexity and misunderstanding have a point.
So I'm removing the *WC funcs and moving ctx into the regular error
generators.  Callers are allowed to provide a nil context if they
don't have one.

The goal here is simplification: don't ask developers to make decisions,
just provide a context when you have one.