move datanode to internal/node
Clues has grown organically into a split between the api that places
data in the right spot (clues) and the structure which contains the data
(nodes).  As we move more behavior into subpackages, and as we look
forward to future subpackage additions, it makes the most sense to
move the node systems into an internal package so that they can be
generically utilized across all clues subpackages without issues
like import cycles.