hook otel support up through clog (#64)
Adds otel support for logging via clog.

Any clog log will automatically duplicate the log via otel logging. We
may want deduplication in the future, which will take more effort to
support a better interface for logger plug-in. That way we could
configure a run for zap, otel, or otel-via-zap.