change file setting semantics (#68)
1. Always default logging to stderr.
2. allow users to overwrite output to stdout or a file. However, to
provide defaults, this setting is unexported and has no hooks for
globally declared files or env varibles. Users must instead pass a file
to the settings struct via a method instead.

This PR is to solve concerns about the capacity for discrete logging
files to appear on the same machine when the intent is for
initialization
to use the same file every time. It requires users to be more aware of
the file locations they're working with in order to set up clog
correctly.