allow noop initiation of stats (#141)
Ctats prints a debug log when a request attempts to update the counts
without having initialized a node in the ctx. This can be useful to
catch in production deployments, where missing initialization can
cripple telemetry gathering. However, in testing where initialization is
not required, the result is extremely chatty and noisesome.

This PR adds a `ctats.InitializeNoop` constructor which allows ctats to
run silently in cases where no initialization is expected, such as unit
testing. The result is a much more quiet set of test outputs.