Fix bug with reinitializing OTEL logger (#152)
Always source the OTEL logger from the context instead of the logger
singleton. This is important since the OTEL logger can have a different
lifetime than the zap logger singleton. Therefore, sourcing the OTEL
logger from the context ensures that if the OTEL logger is closed and
re-initialized elsewhere logs continue to be sent to the configured OTEL
collector.