add clutel span error funcs (#195)
One of the most common ways of interfacing with spans, aside from
creating and closing them, is setting their status to a failure state.
It's a little bit bulky in raw usage, so here's some quick helpers for
it.

Along the way I wanted to begin peicemealing little bits of the error
handing to an internal package, and this was an opportunity to start
with that by relocating the nil iface checker. It isn't an error
specific check at heart, but since we only use it for errors, it still
lives there.